demo_win_pytorch_normal/requirements.txt
mss==9.0.1
numpy==2.0.0
opencv_python==4.9.0.80
Pillow==10.3.0
pyautogui==0.9.54
pydirectinput==1.0.4
pynput==1.7.7
pywin32==306
ultralytics==8.2.15
